[NextDNS] 'domain-list' handles all the list now
