+++ extensions/nextdns/src/components/domain-list.tsx
@@ -1,10 +1,18 @@
 import { List } from "@raycast/api";
 import { DomainListItem, DomainListProps } from "../types/nextdns";
-import { EmptyView } from "../components/empty-view";
-import { ListItem } from "../components/list-item";
-import { PREFERENCES } from "./constants";
+import { EmptyView } from "./empty-view";
+import { ListItem } from "./list-item";
+import { PREFERENCES } from "../libs/constants";
+import { getDomains } from "../libs/api";
+import { ErrorView } from "./error-view";
 
-export const DomainList: React.FC<DomainListProps> = ({ data, type, isLoading, mutate}) => (
+export const DomainList: React.FC<DomainListProps> = ({ type }) => {
+  const { data, isLoading, error, mutate } = getDomains({ type });
+
+  if (error) {
+    return <ErrorView />;
+  }
+  return (
   <List
     isLoading={isLoading}
     searchBarPlaceholder={`Search ${type}list of ${data.profileName} (${PREFERENCES.nextdns_profile_id}`}
@@ -16,4 +24,4 @@ export const DomainList: React.FC<DomainListProps> = ({ data, type, isLoading, m
     {Object.keys(data).length === 0 && <EmptyView title={`No domains in ${type}list`} icon={{ source: "no_view.png" }} />}
     <EmptyView title="No Results" icon={{ source: "no_view.png" }} />
   </List>
-);
+)};
