+++ extensions/nextdns/src/components/empty-view.tsx
@@ -1,6 +1,6 @@
-import { ActionPanel, Icon, ImageLike, List } from "@raycast/api";
+import { ActionPanel, Icon, Image, List } from "@raycast/api";
 
-export function EmptyView(props: { title: string; icon: Icon | ImageLike }) {
+export function EmptyView(props: { title: string; icon: Icon | Image.ImageLike }) {
   const { title, icon } = props;
 
   return <List.EmptyView title={title} icon={icon} actions={<ActionPanel></ActionPanel>} />;
