+++ extensions/nextdns/src/denylist.tsx
@@ -1,16 +1,7 @@
-import { getDomains } from "./libs/api";
-import { ErrorView } from "./components/error-view";
-import { DomainList } from "./libs/domainList";
+import { DomainList } from "./components/domain-list";
 
-export default function Command() {
-  const type = "deny";
-  const { data, isLoading, error, mutate } = getDomains({ type });
-
-  if (error) {
-    return <ErrorView />;
-  }
-
+export default function DenyList() {
   return (
-    <DomainList data={data} type={type} isLoading={isLoading} mutate={mutate} />
+    <DomainList type="deny" />
   );
 }
