+++ extensions/nextdns/src/types/nextdns.ts
@@ -24,10 +24,10 @@ export type Profile = {
 export type Mutate = MutatePromise<{ result: DomainListItem[]; profileName: string; }>;
 
 export interface DomainListProps {
-  data: { result: DomainListItem[]; profileName: string };
+  // data: { result: DomainListItem[]; profileName: string };
   type: "allow" | "deny";
-  isLoading: boolean;
-  mutate: Mutate;
+  // isLoading: boolean;
+  // mutate: Mutate;
 }
 
 export interface DomainSubmitValues {
